+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - Citizens integration (spells will ignore NPC's)
  - Some multi-world fixes to the Camera spell.
  - Fix permission use with wildcards... umm.. I thought the GM handled this??
+ - Added "Magic.wand.destruct", false by default- use if you want to wipe player inventories of wands if they get somewhere they don't have permission to use wands.
 
 ## 2.9.8